+++ nextgen/config/src/toolchain/rust_config.rs
@@ -1,11 +1,12 @@
+use super::bin_config::BinEntry;
 use crate::validate::validate_semver;
 use schematic::Config;
-use serde::Serialize;
 
 /// Docs: https://moonrepo.dev/docs/config/toolchain#rust
-#[derive(Clone, Config, Debug, Serialize)]
+#[derive(Clone, Config, Debug)]
 pub struct RustConfig {
-    pub bins: Vec<String>,
+    #[setting(nested)]
+    pub bins: Vec<BinEntry>,
 
     pub sync_toolchain_config: bool,